+++ bin/obtain.py
@@ -24,8 +24,8 @@ certbot_cmd_template = '''
     --server https://acme-v02.api.letsencrypt.org/directory \
     --manual \
     --manual-public-ip-logging-ok \
-    --manual-auth-hook "python %(manual_path)s --auth --dns %(dns)s" \
-    --manual-cleanup-hook "python %(manual_path)s --cleanup --dns %(dns)s" \
+    --manual-auth-hook "python %(manual_path)s --auth --dns %(dns)s %(challenge_alias)s" \
+    --manual-cleanup-hook "python %(manual_path)s --cleanup --dns %(dns)s %(challenge_alias)s" \
     %(deploy_hook)s \
     %(domains)s
 '''
@@ -39,6 +39,7 @@ def run(args):
     deploy_hook = '--deploy-hook "python ' + deploy_path + '"' if Utils.is_enable_deployment() else ''
     cert_name = '--cert-name ' + args.cert if args.cert else ''
     force_renewal = '--force-renewal' if args.force else ''
+    challenge_alias = '--challenge-alias ' + args.alias if args.alias else ''
 
     certbot_cmd = certbot_cmd_template % {
         'email': Config['base']['email'],
@@ -47,7 +48,8 @@ def run(args):
         'manual_path': manual_path,
         'dns': args.dns,
         'deploy_hook': deploy_hook,
-        'domains': domains
+        'domains': domains,
+        'challenge_alias': challenge_alias
     }
 
     Logger.info('certbot obtain: ' + certbot_cmd)
@@ -61,6 +63,7 @@ def main():
     parser.add_argument('-c', '--cert', help='certificate name, e.g. domain.com')
     parser.add_argument('-f', '--force', help='force renewal', default=False, action='store_true')
     parser.add_argument('--dns', help='dns type, default: aliyun', default='aliyun')
+    parser.add_argument('--challenge-alias', dest='alias', help='challenge aliased domain, e.g. alias.domain.com')
 
     args = parser.parse_args()